refactor: move top level validation into validation files
